add generic source display status support
