+++ src/main/java/org/janelia/saalfeldlab/paintera/state/MinimalSourceState.java
@@ -11,6 +11,7 @@ import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
+import javafx.scene.Node;
 import net.imglib2.converter.Converter;
 import net.imglib2.type.numeric.ARGBType;
 import org.janelia.saalfeldlab.paintera.composition.Composite;
@@ -174,4 +175,9 @@ public class MinimalSourceState<D, T, S extends DataSource<D, T>, C extends Conv
 		return this.isDirtyProperty().get();
 	}
 
+	@Override
+	public Node getDisplayStatus()
+	{
+		return null;
+	}
 }
